requirements.txt
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.16.0
cssselect2==0.7.0
defusedxml==0.7.1
greenlet==3.0.3
imageio==2.34.2
lxml==5.2.2
numpy==2.0.1
pillow==10.4.0
playwright==1.45.1
pycparser==2.22
pyee==11.1.0
scipy==1.14.0
svgpathtools==1.6.1
svgwrite==1.4.3
tinycss2==1.3.0
typing_extensions==4.12.2
webencodings==0.5.1
